+++ docs/changes.rst
@@ -1,6 +1,17 @@
 Changes
 =======
 
+weblate 3.7
+-----------
+
+Released on ? 2019.
+
+* Added separate Celery queue for notifications.
+* Use consistent look with application for API browsing.
+* Include approved stats in the reports.
+* Report progress when updating translation component.
+* Allow to abort running background component update.
+
 weblate 3.6.1
 -------------